Make type coercion ultra-modular
