--- app/resonant-laboratory/general_purpose/binningLogic.js
@@ -1,7 +1,0 @@
-function test (value) {
-  return value + ' there';
-}
-
-var es6exports = { // eslint-disable-line no-unused-vars
-  test: test
-};
